pkgdir/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diskpatrol"
description = "Disk Patrol application"
license = {text = "GPL version 3"}
readme = "README.rst"
requires-python = ">=3.9"
authors = [{name = "Heechul Kim", email = "jijisa@iorchard.net"}]
maintainers = [{name = "Heechul Kim", email = "jijisa@iorchard.net"}]
dependencies = [
    "python-dotenv",
    "requests",
    "chardet"
]
dynamic = ["version"]

[project.scripts]
diskpatrol = "diskpatrol.cli:main"

[project.urls]
"Repository" = "https://github.com/jijisa/diskpatrol.git"

[tool.setuptools.dynamic]
version = {attr = "diskpatrol.__version__"}
